# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/content-analysis/summarize": {
      "post": {
        "description": "Herkes tarafından erişilebilir. LangChain ve Gemini 2.0 Flash ile güçlendirilmiş detaylı eğitimsel özetleme.",
        "operationId": "ContentAnalysisController_summarize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "İçerik başarıyla özetlendi"
          },
          "400": {
            "description": "Geçersiz istek verisi"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        },
        "summary": "AI kullanarak gelişmiş içerik özetleme",
        "tags": [
          "content-analysis"
        ]
      }
    },
    "/content-analysis/summarize-file": {
      "post": {
        "description": "Herkes tarafından erişilebilir. LangChain ve Gemini 2.0 Flash ile güçlendirilmiş dosya özetleme.",
        "operationId": "ContentAnalysisController_summarizeFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeFileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dosya içeriği başarıyla özetlendi"
          },
          "400": {
            "description": "Geçersiz istek verisi"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        },
        "summary": "AI kullanarak gelişmiş dosya içeriği özetleme",
        "tags": [
          "content-analysis"
        ]
      }
    },
    "/content-analysis/generate-quiz-questions": {
      "post": {
        "description": "Herkes tarafından erişilebilir. LangChain ve Gemini 2.0 Flash ile güçlendirilmiş quiz sorusu üretimi.",
        "operationId": "ContentAnalysisController_generateQuizQuestions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuizQuestionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz soruları başarıyla oluşturuldu"
          },
          "400": {
            "description": "Geçersiz istek verisi"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        },
        "summary": "LangChain ile gelişmiş quiz soruları oluşturma",
        "tags": [
          "content-analysis"
        ]
      }
    },
    "/content-analysis/analyze-workflow": {
      "post": {
        "description": "Herkes tarafından erişilebilir. LangGraph workflow ve Gemini 2.0 Flash ile güçlendirilmiş içerik analizi.",
        "operationId": "ContentAnalysisController_analyzeContentWorkflow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeContentWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "İçerik analizi başarıyla tamamlandı"
          },
          "400": {
            "description": "Geçersiz istek verisi"
          },
          "500": {
            "description": "Sunucu hatası"
          }
        },
        "summary": "LangGraph workflow ile gelişmiş içerik analizi",
        "tags": [
          "content-analysis"
        ]
      }
    },
...
    "/dashboard/weaknesses-analysis": {
      "get": {
        "operationId": "DashboardController_getWeaknessesAnalysis",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Eksiklik analizi başarıyla alındı"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcının eksiklik analizi ve iyileştirme önerileri",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/progress-tracking": {
      "get": {
        "operationId": "DashboardController_getProgressTracking",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "İlerleme takibi başarıyla alındı"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcının ilerleme takibi ve istatistikleri",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/study-recommendations": {
      "get": {
        "operationId": "DashboardController_getStudyRecommendations",
        "parameters": [
          {
            "name": "focus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Çalışma önerileri başarıyla oluşturuldu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kişiselleştirilmiş çalışma önerileri",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/achievement-summary": {
      "get": {
        "operationId": "DashboardController_getAchievementSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Başarı özeti başarıyla alındı"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kullanıcının başarı özeti ve motivasyon verileri",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/dashboard/personalized-homepage": {
      "get": {
        "operationId": "DashboardController_getPersonalizedHomepage",
        "parameters": [
          {
            "name": "includeProgress",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "lessonCount",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ana sayfa içeriği başarıyla oluşturuldu"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kişiselleştirilmiş ana sayfa içeriği",
        "tags": [
          "Dashboard"
        ]
      }
    }
  },
  "info": {
    "title": "🎓 GÜNÜBİRLİK DOZ API",
...
    "version": "2.0.0",
    "contact": {
      "name": "GÜNÜBİRLİK DOZ Geliştirici Ekibi",
      "url": "https://github.com/gunibirlikdoz",
      "email": "info@gunibirlikdoz.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT Bearer token ile kimlik doğrulama",
        "in": "header"
      }
    },
    "schemas": {
...
      "SummarizeContentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text content to summarize"
          },
          "videoUrl": {
            "type": "string",
            "description": "URL of the video (optional)"
          },
          "title": {
            "type": "string",
            "description": "Title of the content (optional)"
          },
          "subject": {
            "type": "string",
            "description": "Subject area (e.g., Mathematics, Science, History)"
          },
          "gradeLevel": {
            "type": "string",
            "description": "Grade level (e.g., 5th Grade, 8th Grade, High School)"
          },
          "learningObjectives": {
            "type": "string",
            "description": "Learning objectives or goals"
          },
          "targetAudience": {
            "type": "string",
            "description": "Target audience (e.g., students, teachers, parents)"
          },
          "difficultyLevel": {
            "type": "string",
            "description": "Difficulty level (beginner, intermediate, advanced)"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration of the content in minutes"
          },
          "keyTopics": {
            "type": "string",
            "description": "Key topics or keywords to focus on"
          },
          "summaryType": {
            "type": "string",
            "description": "Type of summary needed (brief, detailed, educational)",
            "default": "educational"
          }
        },
        "required": [
          "text"
        ]
      },
      "SummarizeFileDto": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "description": "Path to the file to summarize"
          },
          "title": {
            "type": "string",
            "description": "Title of the content (optional)"
          },
          "subject": {
            "type": "string",
            "description": "Subject area (e.g., Mathematics, Science, History)"
          },
          "gradeLevel": {
            "type": "string",
            "description": "Grade level (e.g., 5th Grade, 8th Grade, High School)"
          },
          "learningObjectives": {
            "type": "string",
            "description": "Learning objectives or goals"
          },
          "targetAudience": {
            "type": "string",
            "description": "Target audience (e.g., students, teachers, parents)"
          },
          "difficultyLevel": {
            "type": "string",
            "description": "Difficulty level (beginner, intermediate, advanced)"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration of the content in minutes"
          },
          "keyTopics": {
            "type": "string",
            "description": "Key topics or keywords to focus on"
          },
          "summaryType": {
            "type": "string",
            "description": "Type of summary needed (brief, detailed, educational)",
            "default": "educational"
          }
        },
        "required": [
          "filePath"
        ]
      },
      "GenerateQuizQuestionsDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text content to generate questions from"
          },
          "numberOfQuestions": {
            "type": "number",
            "description": "Number of questions to generate",
            "default": 5
          },
          "subject": {
            "type": "string",
            "description": "Subject area (e.g., Mathematics, Science, History)"
          },
          "gradeLevel": {
            "type": "string",
            "description": "Grade level (e.g., 5th Grade, 8th Grade, High School)"
          },
          "difficultyLevel": {
            "type": "string",
            "description": "Difficulty level (easy, medium, hard)",
            "default": "medium"
          },
          "questionType": {
            "type": "string",
            "description": "Question type (multiple-choice, true-false, fill-blank)",
            "default": "multiple-choice"
          },
          "learningObjectives": {
            "type": "string",
            "description": "Learning objectives to focus on"
          },
          "keyTopics": {
            "type": "string",
            "description": "Key topics or keywords to emphasize"
          },
          "language": {
            "type": "string",
            "description": "Language for questions (Turkish, English)",
            "default": "Turkish"
          }
        },
        "required": [
          "text"
        ]
      },
      "AnalyzeContentWorkflowDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text content to analyze with LangGraph workflow"
          },
          "analysisType": {
            "type": "string",
            "description": "Type of analysis to perform",
            "enum": [
              "summary",
              "detailed",
              "educational"
            ],
            "default": "summary"
          },
          "subject": {
            "type": "string",
            "description": "Subject area (e.g., Mathematics, Science, History)"
          },
          "gradeLevel": {
            "type": "string",
            "description": "Grade level (e.g., 5th Grade, 8th Grade, High School)"
          },
          "learningObjectives": {
            "type": "string",
            "description": "Learning objectives or goals"
          },
          "targetAudience": {
            "type": "string",
            "description": "Target audience (e.g., students, teachers, parents)"
          },
          "difficultyLevel": {
            "type": "string",
            "description": "Difficulty level (beginner, intermediate, advanced)"
          },
          "keyTopics": {
            "type": "string",
            "description": "Key topics or keywords to focus on"
          },
          "analysisDepth": {
            "type": "string",
            "description": "Analysis depth (shallow, moderate, deep)",
            "default": "moderate"
          },
          "includeRecommendations": {
            "type": "boolean",
            "description": "Include educational recommendations",
            "default": true
          },
          "language": {
            "type": "string",
            "description": "Language for analysis output (Turkish, English)",
            "default": "Turkish"
          }
        },
        "required": [
          "text"
        ]
      },
...
    }
  },
...
}